+++ e2e/wdio/platform/fixtures/appData/action-list-item-contents.ts
@@ -1,5 +1,4 @@
-export default {
-    btnText: [
+export const btnText = [
         'Action 1',
         'Action 2',
         'Action 3',
@@ -16,5 +15,4 @@ export default {
         'Action 2',
         'Action 3',
         'Action 4'
-    ]
-};
+];