+++ e2e/wdio/platform/fixtures/appData/input-page-contents.ts
@@ -1,25 +1,23 @@
-export default {
-    labelsArray: [
-        'Default Input Field',
-        'Text Input Field',
-        'Number Input Field',
-        'Compact Input Field',
-        'ReadOnly Input Field',
-        'Disabled Input Field',
-        'Inline Help Input Field'
-    ],
-    favoriteColor: 'My',
-    placeholdersArray: [
-        'Field placeholder text',
-        'Field placeholder text',
-        'Field placeholder text',
-        'Field placeholder text',
-        'Field placeholder text',
-        'Field placeholder text',
-        'Field placeholder text',
-        'Field placeholder text',
-        'Enter the sport name'
-    ],
-    errorText: 'Pellentesque metus lacus commodo eget justo ut rutrum varius nunc',
-    errorBorderColor: '187,0,0'
-};
+export const labelsArray = [
+    'Default Input Field',
+    'Text Input Field',
+    'Number Input Field',
+    'Compact Input Field',
+    'ReadOnly Input Field',
+    'Disabled Input Field',
+    'Inline Help Input Field'
+];
+export const favoriteColor = 'My';
+export const placeholdersArray = [
+    'Field placeholder text',
+    'Field placeholder text',
+    'Field placeholder text',
+    'Field placeholder text',
+    'Field placeholder text',
+    'Field placeholder text',
+    'Field placeholder text',
+    'Field placeholder text',
+    'Enter the sport name'
+];
+export const errorText = 'Pellentesque metus lacus commodo eget justo ut rutrum varius nunc';
+export const errorBorderColor = '187,0,0';