+++ e2e/wdio/platform/fixtures/appData/panel-page-content.ts
@@ -1,5 +1,3 @@
-export default {
-    expandable_panel_header: 'Panel Title',
-    action_panel_edit_button: 'Edit',
-    action_panel_delete_button: 'Delete'
-};
+export const expandable_panel_header = 'Panel Title';
+export const action_panel_edit_button = 'Edit';
+export const action_panel_delete_button = 'Delete';