+++ e2e/wdio/platform/tests/object-list-item.e2e-spec.ts
@@ -11,7 +11,7 @@ import {
 } from '../../driver/wdio';
 import { ObjectListItemPo } from '../pages/object-list-item.po';
 import { checkElArrIsClickable, checkElementDisplayed, checkElementText } from '../../helper/assertion-helper';
-import ObjListData from '../fixtures/appData/object-list-item-contents';
+import {noBorderAttr, navUrl, altNoBorderStyle, noBorderStyle} from '../fixtures/appData/object-list-item-contents';
 
 describe('Object list item suite:', function() {
     const objListPg = new ObjectListItemPo();
@@ -40,14 +40,14 @@ describe('Object list item suite:', function() {
 
     describe('Object List Item examples:', function() {
         it('should check styles and content', () => {
-           expect(getAttributeByName(objListPg.objListAttr, ObjListData.noBorderAttr)).toBe('true');
+           expect(getAttributeByName(objListPg.objListAttr, noBorderAttr)).toBe('true');
            checkElementDisplayed(objListPg.obJListIntro);
            checkElementDisplayed(objListPg.objListAttributes);
            checkElementDisplayed(objListPg.objListStatuses);
            if (browserIsFirefox()) {
-               expect(getCSSPropertyByName(objListPg.objListItem, ObjListData.altNoBorderStyle).value).toBe('none');
+               expect(getCSSPropertyByName(objListPg.objListItem, altNoBorderStyle).value).toBe('none');
            } else {
-               expect(getCSSPropertyByName(objListPg.objListItem, ObjListData.noBorderStyle).value).toBe('none');
+               expect(getCSSPropertyByName(objListPg.objListItem, noBorderStyle).value).toBe('none');
            }
         });
     });
@@ -80,7 +80,7 @@ describe('Object list item suite:', function() {
         it('should check navigation', () => {
             click(objListPg.objNavList);
             const currentUrl = getCurrentUrl();
-            expect(currentUrl).toContain(ObjListData.navUrl);
+            expect(currentUrl).toContain(navUrl);
             objListPg.open();
         });
     });