+++ wdio.conf.js
@@ -266,7 +266,7 @@ exports.config = {
     jasmineNodeOpts: {
         isVerbose: true,
         showColors: true,
-        defaultTimeoutInterval: 700000,
+        defaultTimeoutInterval: 1200000,
         grep: null,
         invertGrep: null
     },